Re-integrate RustPython parser repository (#4359)

+++ crates/ruff/src/rules/isort/types.rs
@@ -14,7 +14,7 @@ pub enum TrailingComma {
 #[derive(Debug, Hash, Ord, PartialOrd, Eq, PartialEq, Clone)]
 pub struct ImportFromData<'a> {
     pub module: Option<&'a str>,
-    pub level: Option<usize>,
+    pub level: Option<u32>,
 }
 
 #[derive(Debug, Hash, Ord, PartialOrd, Eq, PartialEq)]
